Completely updated for v1.6.0
- Removed ClassicWhiteBorder/ClassicGoldBorder (this is now a settings feature)
- Removed UniversesBeyond (now a base template)
- Updated directory and pathing conventions
