+++ .gitignore
@@ -1,0 +1,3 @@
+{\rtf1}
+*.psd
+/configs/*.ini
