--- proxyshop/plugins/WarpDandy/configs/SamuraiTemplate.ini
@@ -1,28 +1,0 @@
-[FILES]
-Output.Filetype = jpg
-Save.Artist.Name = 1
-Overwrite.Duplicate = 1
-
-[TEXT]
-Language = en
-Flavor.Divider = 1
-No.Flavor.Text = 0
-No.Reminder.Text = 0
-True.Collector.Info = 1
-Force.English.Formatting = 0
-
-[SYMBOLS]
-Default.Symbol = 
-Auto.Set.Symbol = 1
-Auto.Symbol.Size = 1
-Symbol.Stroke.Size = 6
-Fill.Symbol.Background = 0
-
-[APP]
-Manual.Edit = 0
-Skip.Failed = 0
-Scryfall.Ascending = 0
-Targeted.Replace = 1
-Render.Snow = 0
-Dev.Mode = 0
-
