--- proxyshop/plugins/WarpDandy/template_map.json
@@ -1,33 +1,0 @@
-{
-	"normal": {
-		"Samurai": "SamuraiTemplate",
-		"Ninja": "NinjaTemplate",
-		"Ninja Glitch": "NinjaGlitchTemplate",
-		"Mirrored": "MirroredTemplate",
-		"Classic White Border": "ClassicWhiteBorderTemplate",
-		"Nickname Small": "NicknameSmallTemplate",
-		"Nickname Medium": "NicknameMediumTemplate",
-		"Nickname Large": "NicknameLargeTemplate",
-		"Nickname": "NicknameBaseTemplate",
-		"Golden Age": "GoldenAgeTemplate",
-		"Original Expedition": "zneExpeditionTemplate",
-		"Skyscraper": "SkyscraperTemplate",
-		"Golden Age V2": "GoldenAgeV2Template",
-		"Art Deco": "ArtDecoTemplate",
-		"Universes Beyond": "UniversesBeyondTemplate",
-		"Etched": "EtchedTemplate",
-		"Destiny": "DestinyTemplate",
-		"Neon Extended": "PinlinesExtNeonTemplate",
-		"Mystical Archive": "MysticalArchiveTemplate",
-		"Crimson Fang Extended": "FangExtendedTemplate",
-		"Golden Age Full Art": "GoldenAgeFullArtTemplate",
-		"Classic Gold Border": "ClassicGoldBorderTemplate",
-		"No Border": "NoBorderTemplate"
-		
-	},
-	"planeswalker": {
-		"Classic PW": "ClassicPWTemplate",
-		"Art Deco PW": "ArtDecoPWTemplate"
-	
-	}
-}
