+++ template_map.json
@@ -1,0 +1,28 @@
+{
+	"normal": {
+		"Samurai": "SamuraiTemplate",
+		"Ninja": "NinjaTemplate",
+		"Ninja Glitch": "NinjaGlitchTemplate",
+		"Mirrored": "MirroredTemplate",
+		"Nickname Small": "NicknameSmallTemplate",
+		"Nickname Medium": "NicknameMediumTemplate",
+		"Nickname Large": "NicknameLargeTemplate",
+		"Nickname": "NicknameBaseTemplate",
+		"Golden Age": "GoldenAgeTemplate",
+		"Original Expedition": "zneExpeditionTemplate",
+		"Skyscraper": "SkyscraperTemplate",
+		"Golden Age V2": "GoldenAgeV2Template",
+		"Art Deco": "ArtDecoTemplate",
+		"Etched": "EtchedTemplate",
+		"Destiny": "DestinyTemplate",
+		"Neon Extended": "PinlinesExtNeonTemplate",
+		"Mystical Archive": "MysticalArchiveTemplate",
+		"Crimson Fang Extended": "FangExtendedTemplate",
+		"Golden Age Full Art": "GoldenAgeFullArtTemplate",
+		"No Border": "NoBorderTemplate"
+	},
+	"planeswalker": {
+		"Classic PW": "ClassicPWTemplate",
+		"Art Deco PW": "ArtDecoPWTemplate"
+	}
+}
